Bare bones Start Menu Up and running
To use as is, when building the scenes, set mine as the first scene, and the gym as the next scene. When Start Game is selected, it will open the gym scene.

If you want to change the scene to lead when Start Game button is pressed, change the name in the following line:
SceneManager.LoadScene("Gym");

Where "Gym" is the scene to load